ListItem: * Switched all checkboxes to be green * Gave IncludeExcludeFilter listItems correct UI * compiled grocery & kitchen list into provisionsList * fixed incorrect call for trash image * Removed old kitchenList & myKitchenItem files
