+++ Food@Home/ReuseableViews/RecipeDetails/RecipeDetailsView/Components/MealScheduler/MealScheduler.swift
@@ -78,6 +78,7 @@ struct MealScheduler: View {
                             timeSelectors.append(TimeSelectorObject(date: nextAvaliableDate()))
                         }
                         .disabled(timeSelectors.count > 14)
+                        .tint(.blue)
                     }
                 }
                 .padding(.horizontal, 17)
